filter package name renamed to filters
