+++ app/src/main/java/com/trx/consumer/screens/discover/filters/FiltersAdapter.kt
@@ -1,4 +1,4 @@
-package com.trx.consumer.screens.discover.filter.filteroptions
+package com.trx.consumer.screens.discover.filters
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -6,15 +6,15 @@ import com.trx.consumer.R
 import com.trx.consumer.common.CommonRecyclerViewAdapter
 import com.trx.consumer.common.CommonViewHolder
 import com.trx.consumer.managers.LogManager
-import com.trx.consumer.models.common.FilterOptionsModel
+import com.trx.consumer.models.common.FilterModel
 import com.trx.consumer.views.EmptyViewHolder
 import kotlinx.coroutines.CoroutineScope
 
-class FilterOptionsAdapter(
-    private val listener: FilterOptionsListener,
+class FiltersAdapter(
+    private val listener: FiltersListener,
     scopeProvider: () -> CoroutineScope
-) :
-    CommonRecyclerViewAdapter(scopeProvider) {
+) : CommonRecyclerViewAdapter(scopeProvider) {
+
     companion object {
         private const val TYPE_EMPTY = -1
         private const val TYPE_ROW = 1
@@ -25,9 +25,9 @@ class FilterOptionsAdapter(
     override fun createCommonViewHolder(parent: ViewGroup, viewType: Int): CommonViewHolder {
         return try {
             when (viewType) {
-                TYPE_ROW -> FilterOptionsViewHolder(
+                TYPE_ROW -> FiltersViewHolder(
                     LayoutInflater.from(parent.context)
-                        .inflate(R.layout.row_filter_options, parent, false)
+                        .inflate(R.layout.row_filters, parent, false)
                 )
                 else -> EmptyViewHolder(
                     LayoutInflater.from(parent.context).inflate(R.layout.row_empty, parent, false)
@@ -44,21 +44,25 @@ class FilterOptionsAdapter(
     override fun onBindViewHolder(holder: CommonViewHolder, position: Int) {
         val item = items[position]
         when (holder) {
-            is FilterOptionsViewHolder -> {
-                holder.setup(item as FilterOptionsModel, listener)
+            is FiltersViewHolder -> {
+                holder.setup(item as FilterModel, listener)
             }
         }
     }
 
-    fun update(newItems: List<FilterOptionsModel>) {
+    fun update(newItems: List<FilterModel>) {
         this.items.clear()
         this.items.addAll(newItems)
         this.notifyDataSetChanged()
     }
 
+    fun reset() {
+        // TODO:
+    }
+
     override fun getItemViewType(position: Int): Int {
         return when (items[position]) {
-            is FilterOptionsModel -> TYPE_ROW
+            is FilterModel -> TYPE_ROW
             else -> TYPE_EMPTY
         }
     }
