+++ app/src/main/java/com/trx/consumer/screens/discover/filters/filteroptions/FilterOptionsAdapter.kt
@@ -1,4 +1,4 @@
-package com.trx.consumer.screens.discover.filter
+package com.trx.consumer.screens.discover.filters.filteroptions
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -6,15 +6,15 @@ import com.trx.consumer.R
 import com.trx.consumer.common.CommonRecyclerViewAdapter
 import com.trx.consumer.common.CommonViewHolder
 import com.trx.consumer.managers.LogManager
-import com.trx.consumer.models.common.FilterModel
+import com.trx.consumer.models.common.FilterOptionsModel
 import com.trx.consumer.views.EmptyViewHolder
 import kotlinx.coroutines.CoroutineScope
 
-class FiltersAdapter(
-    private val listener: FiltersListener,
+class FilterOptionsAdapter(
+    private val listener: FilterOptionsListener,
     scopeProvider: () -> CoroutineScope
-) : CommonRecyclerViewAdapter(scopeProvider) {
-
+) :
+    CommonRecyclerViewAdapter(scopeProvider) {
     companion object {
         private const val TYPE_EMPTY = -1
         private const val TYPE_ROW = 1
@@ -25,9 +25,9 @@ class FiltersAdapter(
     override fun createCommonViewHolder(parent: ViewGroup, viewType: Int): CommonViewHolder {
         return try {
             when (viewType) {
-                TYPE_ROW -> FiltersViewHolder(
+                TYPE_ROW -> FilterOptionsViewHolder(
                     LayoutInflater.from(parent.context)
-                        .inflate(R.layout.row_filters, parent, false)
+                        .inflate(R.layout.row_filter_options, parent, false)
                 )
                 else -> EmptyViewHolder(
                     LayoutInflater.from(parent.context).inflate(R.layout.row_empty, parent, false)
@@ -44,25 +44,21 @@ class FiltersAdapter(
     override fun onBindViewHolder(holder: CommonViewHolder, position: Int) {
         val item = items[position]
         when (holder) {
-            is FiltersViewHolder -> {
-                holder.setup(item as FilterModel, listener)
+            is FilterOptionsViewHolder -> {
+                holder.setup(item as FilterOptionsModel, listener)
             }
         }
     }
 
-    fun update(newItems: List<FilterModel>) {
+    fun update(newItems: List<FilterOptionsModel>) {
         this.items.clear()
         this.items.addAll(newItems)
         this.notifyDataSetChanged()
     }
 
-    fun reset() {
-        // TODO:
-    }
-
     override fun getItemViewType(position: Int): Int {
         return when (items[position]) {
-            is FilterModel -> TYPE_ROW
+            is FilterOptionsModel -> TYPE_ROW
             else -> TYPE_EMPTY
         }
     }
