+++ app/src/main/java/com/trx/consumer/screens/discover/filters/filteroptions/FilterOptionsListener.kt
@@ -1,4 +1,4 @@
-package com.trx.consumer.screens.discover.filter.filteroptions
+package com.trx.consumer.screens.discover.filters.filteroptions
 
 import com.trx.consumer.models.common.FilterOptionsModel
