+++ app/src/main/res/navigation/nav_main.xml
@@ -104,7 +104,7 @@
 
     <fragment
         android:id="@+id/filter_options_fragment"
-        android:name="com.trx.consumer.screens.discover.filter.filteroptions.FilterOptionsFragment"
+        android:name="com.trx.consumer.screens.discover.filters.filteroptions.FilterOptionsFragment"
         android:label="FilterOptionsFragment"
         tools:layout="@layout/fragment_filter_options" />
 
@@ -116,7 +116,7 @@
 
     <fragment
         android:id="@+id/filter_fragment"
-        android:name="com.trx.consumer.screens.discover.filter.FiltersFragment"
+        android:name="com.trx.consumer.screens.discover.filters.FiltersFragment"
         android:label="FilterFragment"
         tools:layout="@layout/fragment_filters" />
